changed toolset to clang & fixed bugs
