+++ SPS Communication/SPS Communication.vcxproj.filters
@@ -71,5 +71,8 @@
     <ClInclude Include="VariableSequence.h">
       <Filter>Header Files\Processing</Filter>
     </ClInclude>
+    <ClInclude Include="SPSIO.h">
+      <Filter>Header Files\SPS Connection</Filter>
+    </ClInclude>
   </ItemGroup>
 </Project>
