+++ SPS Communication/StateElement.h
@@ -1,16 +1,17 @@
 #pragma once
 
 #include "Includes.h"
+#include "Logging.h"
 #include "Parser.h"
 #include "VariableSequence.h"
 
 template<typename _VarSeq, template<typename> class... Ex>
-class basic_StateElement : std::vector<_VarSeq>, public Ex<basic_StateElement<Ex...>>...
+class basic_StateElement : std::vector<_VarSeq>, public Ex<basic_StateElement<_VarSeq, Ex...>>...
 {
 	using vec_t = std::vector<_VarSeq>;
 
 public:
-	static constexpr std::string_view VAR_NAME = "[state]=>"
+	static constexpr std::string_view VAR_NAME = "[state]=>";
 
 	using varseq_t = _VarSeq;
 
@@ -47,8 +48,9 @@ public:
 		{
 			underlying()->emplace_back();
 
-			if (const auto str = p.get_until({ '|', '\n' }); str.has_value())
-				underlying()->back().parse(str.value());
+			if (const auto str = p.get_until(std::array{ '|', '\n' }); str.has_value())
+				underlying()->back().parse(str.value()),
+				p.mov(1);
 			else
 				throw Logger("Couldn't read state message.");
 		}
