+++ SPS Communication/main.cpp
@@ -1,9 +1,11 @@
 #include "Includes.h"
+#include "Logging.h"
 #include "Server.h"
 #include "SPS.h"
 #include "Message.h"
 #include "Request.h"
 #include "AuthElement.h"
+#include "StateElement.h"
 
 std::string get_auth_code(asio::io_context& io)
 {
@@ -16,12 +18,14 @@ std::string get_auth_code(asio::io_context& io)
 		}
 		catch (const std::exception & e)
 		{
-			g_log.write("Failed getting authentication code. ");
+			g_log.write("Failed getting authentication code. Error: " + std::string(e.what()) + '\n');
 			if (test_case == 2)
 				throw Logger("");
 			g_log.write(std::string("Trying again... ") + static_cast<char>(test_case + '0' + 1) + " of 3.\n");
 		}
 	}
+
+	return "";
 }
 
 
@@ -36,7 +40,7 @@ int main(int argc, char** argv)
 		return 1;
 	}
 
-	//SPS sps;
+	SPS sps;
 	//sps.connect(argv[1]);
 
 	g_log.seperate();
@@ -49,6 +53,7 @@ int main(int argc, char** argv)
 	constexpr std::string_view message =
 		"#START\n"
 		"#DEBUG\n"
+		"Hello There!\n"
 		"#DATA\n"
 		"[requesttimeout]=>5\n"
 		"[state]=>0!0_1!1_89!|1!0_1!2_100!\n"
@@ -57,6 +62,7 @@ int main(int argc, char** argv)
 	constexpr std::string_view message_auth =
 		"#START\n"
 		"#DEBUG\n"
+		"Hello There!\n"
 		"#DATA\n"
 		"[requesttimeout]=>5\n"
 		"[authcode]=>123456789asdfghjkl\n"
@@ -64,7 +70,7 @@ int main(int argc, char** argv)
 
 	//Filter through authcode
 	RequestProcessing curr_req;
-	curr_req.parse(message);
+	curr_req.parse(message_auth);
 	curr_req.print_debug();
 
 	CommandList<AuthElement> auth_list;
@@ -78,7 +84,12 @@ int main(int argc, char** argv)
 	{
 		std::this_thread::sleep_until(timeout);
 
+		RequestProcessing req;
+		req.parse(message);
+		req.print_debug();
 
+		CommandList<StateElement> state_list;
+		state_list.parse(req.data());
 
 		////Get data from SPS
 		//const auto& curr_com = com.get();
